Initial commit of fork
Lots of changes in my personal firmware...

Printrboard Jr. config
Support for SainSmart 2004 I2C display
Lots of changes to LCD status screens (progress indicators, etc)
Device metrics saved in EEPROM
Additional M codes supported

+++ Marlin/cardreader.h
@@ -38,7 +38,7 @@ public:
   FORCE_INLINE bool eof() { return sdpos>=filesize ;};
   FORCE_INLINE int16_t get() {  sdpos = file.curPosition();return (int16_t)file.read();};
   FORCE_INLINE void setIndex(long index) {sdpos = index;file.seekSet(index);};
-  FORCE_INLINE uint8_t percentDone(){if(!sdprinting) return 0; if(filesize) return sdpos*100/filesize; else return 0;};
+  FORCE_INLINE float percentDone(){if(!sdprinting) return 0.0; if(filesize) return sdpos*100.0/(float)filesize; else return 0.0;};
   FORCE_INLINE char* getWorkDirName(){workDir.getFilename(filename);return filename;};
 
 public:
